server/requirements_BACKUP_POST_LOCKDOWN_STABLE.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.3.2
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.109.0
frozenlist==1.6.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
passlib==1.7.4
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.13.2
uvicorn==0.27.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
